Доводка экрана поиска
