+++ lib/ui/screen/sight_search_screen.dart
@@ -8,19 +8,17 @@ import 'package:places/ui/const/app_icons.dart';
 import 'package:places/ui/const/app_strings.dart';
 import 'package:places/ui/screen/res/theme_extension.dart';
 import 'package:places/ui/screen/sight_details.dart';
-import 'package:places/ui/widget/buttons/svg_text_button.dart';
-import 'package:places/ui/widget/common.dart';
 import 'package:places/ui/widget/darken_image.dart';
 import 'package:places/ui/widget/empty_list.dart';
 import 'package:places/ui/widget/loader.dart';
 import 'package:places/ui/widget/search_bar.dart';
+import 'package:places/ui/widget/search_history.dart';
 import 'package:places/ui/widget/simple_app_bar.dart';
-import 'package:places/ui/widget/svg_icon.dart';
 
 /// Экран "Поиск мест" по названию.
 ///
 /// Содержимое зависит от состояния [_SightSearchScreenState._state] типа [SearchState].
-/// При входе на экран отображается история поиска [_HistoryWidget],
+/// При входе на экран отображается история поиска [SearchHistory],
 /// она же отображается при очистке поля ввода.
 /// Для возврата на предыдущий экран необходимо нажать иконку "Очистить" при пустом поле ввода.
 class SightSearchScreen extends StatefulWidget {
@@ -71,7 +69,7 @@ class _SightSearchScreenState extends State<SightSearchScreen> {
           builder: (context, state, _) {
             switch (state) {
               case SearchState.initial:
-                return _HistoryWidget(
+                return SearchHistory(
                   historyProvider: _historyProvider,
                   onItemTap: (text) {
                     _controller
@@ -175,14 +173,8 @@ class _SightSearchScreenState extends State<SightSearchScreen> {
             growable: false,
           );
 
-      // Проверка индикатора прогресса
-      // await Future<void>.delayed(const Duration(seconds: 1));
-
-      // Проверка обработки ошибок
-      // throw Exception('testException');
-
-      // Если за время поиска успели очистить поле ввода, то переходим в initial,
-      // иначе показываем результат
+      // Если за время поиска успели очистить поле ввода,
+      // значит результат уже не интересен, отображаем историю
       _state.value = _controller.text.isEmpty
           ? SearchState.initial
           : _filtered.isEmpty
@@ -192,103 +184,6 @@ class _SightSearchScreenState extends State<SightSearchScreen> {
       Utils.log(e.toString());
       _state.value = SearchState.error;
     }
-  }
-}
-
-/// Виджет для отображения истории поиска.
-class _HistoryWidget extends StatelessWidget {
-  final SearchHistoryProvider historyProvider;
-  final void Function(String) onItemTap;
-
-  const _HistoryWidget({
-    Key? key,
-    required this.historyProvider,
-    required this.onItemTap,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return FutureBuilder<Iterable<String>>(
-      future: historyProvider.items(),
-      builder: (_, snapshot) {
-        // Прогресс
-        if (snapshot.connectionState != ConnectionState.done) {
-          return const Loader();
-        }
-
-        // Пустой экран
-        if (snapshot.hasError || !snapshot.hasData || snapshot.data!.isEmpty) {
-          return const SizedBox();
-        }
-
-        // Результат
-        return AnimatedBuilder(
-          animation: historyProvider,
-          builder: (context, child) {
-            if (snapshot.data!.isEmpty) {
-              return const SizedBox();
-            }
-
-            return Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0),
-              child: LayoutBuilder(
-                builder: (context, constraints) {
-                  return Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      spacerH24,
-                      Text(
-                        AppStrings.yourHistory,
-                        style: theme.superSmallInactiveBlack,
-                      ),
-                      ConstrainedBox(
-                        constraints: BoxConstraints(
-                          maxHeight: constraints.maxHeight - 100.0,
-                        ),
-                        // TODO(novikov): ListView.separated
-                        child: SingleChildScrollView(
-                          child: Column(
-                            children: ListTile.divideTiles(
-                              tiles: snapshot.data!.map((e) => ListTile(
-                                    contentPadding: EdgeInsets.zero,
-                                    title: Text(
-                                      e,
-                                      style: theme.text400Secondary2,
-                                    ),
-                                    trailing: IconButton(
-                                      splashRadius: 20.0,
-                                      icon: const SvgIcon(AppIcons.close),
-                                      onPressed: () =>
-                                          historyProvider.remove(e),
-                                    ),
-                                    onTap: () => onItemTap(e),
-                                  )),
-                              color: theme.dividerColor,
-                            ).toList(growable: false),
-                          ),
-                        ),
-                      ),
-                      child!,
-                    ],
-                  );
-                },
-              ),
-            );
-          },
-          child: Transform.translate(
-            offset: const Offset(-8.0, 0.0),
-            child: SvgTextButton.link(
-              padding: const EdgeInsets.symmetric(horizontal: 8.0),
-              label: AppStrings.clearHistory,
-              color: theme.colorScheme.green,
-              onPressed: historyProvider.clear,
-            ),
-          ),
-        );
-      },
-    );
   }
 }
